@@ -4,12 +4,11 @@ from tkinter import ttk
 from typing import *
 
 import json5 as json
-from pygments import styles
 
+from src.Components.tktext import EnhancedTextFrame
+from src.Components.winframe import WinFrame
 from src.types import Tk_Win
 from src.Utils.images import get_image
-from src.Widgets.tktext import EnhancedTextFrame
-from src.Widgets.winframe import WinFrame
 
 
 # Need these to prevent circular imports
