@@ -2,8 +2,8 @@ from tkinter import ttk
 
 import json5 as json
 
+from src.Components.winframe import WinFrame
 from src.Utils.images import get_image
-from src.Widgets.winframe import WinFrame
 
 
 class PluginView(WinFrame):
